refactor: 메뉴그룹 테스트 패키지 위치 수정
